CRUD for Rifa, Premio and User. Relation between Rifa and Premio not working properly

+++ rifaria-api/src/main/java/com/rifaria/rifariaapi/model/Premio.java
@@ -18,22 +18,19 @@ public class Premio implements Serializable {
 	@Id
 	@GeneratedValue
 	@Column(name = "id", updatable = false)
-	private Long id;
+	private Integer id;
 	
 	@Column(name = "titulo")
 	private String titulo;
 	
 	@Column(name = "valor")
 	private double valor;
-	
-//	@OneToOne(mappedBy = "premio")
-//	private Rifa rifa;
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
